requirements.txt
# requirements.txt reset to empty
requests
python-dotenv
